// src/notifications/types/mailchimp.interface.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//

export interface TemplateContent {
  name: string;
  content: string;
}

export interface Metadata {
  website: string;
}

export interface To {
  email: string;
  type: string;
}

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface Headers {}

export interface Message {
  html: string;
  text: string;
  subject: string;
  from_email: string;
  from_name: string;
  to: To[];
  headers: Headers;
  important: boolean;
  track_opens: boolean;
  track_clicks: boolean;
  auto_text: boolean;
  auto_html: boolean;
  inline_css: boolean;
  url_strip_qs: boolean;
  preserve_recipients: boolean;
  view_content_link: boolean;
  bcc_address: string;
  tracking_domain: string;
  signing_domain: string;
  return_path_domain: string;
  merge: boolean;
  merge_language: string;
  global_merge_vars: TemplateContent[];
  merge_vars: TemplateContent[];
  tags: any[];
  subaccount: string;
  google_analytics_domains: any[];
  google_analytics_campaign: string;
  metadata: Metadata;
  recipient_metadata: any[];
  attachments: any[];
  images: any[];
}

export interface MailChimpInterface {
  key: string;
  template_name: string;
  template_content: TemplateContent[];
  message: Message;
  async: boolean;
  ip_pool: string;
  send_at: string;
}